fix: repair CI, Python packaging, and Node 24 compatibility
- Replace failing auto-commit workflow with a proper build job on forks
- Add the missing social-analyzer Python package for pip install -e .
- Update Dockerfile to Node 24, fix Node version check, and add npm test
- Make test/test.sh validate the Python module via --list
