+++ package.json
@@ -4,10 +4,10 @@
   "description": "Advanced Social Media Profiles Finder and String Analysis Tool",
   "main": "app.js",
   "engines": {
-    "node": ">=20.18.1"
+    "node": ">=24"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
+    "test": "node --check app.js && node app.js --help",
     "start": "node app.js --gui"
   },
   "dependencies": {
@@ -25,7 +25,7 @@
     "https-proxy-agent": "5.0.1",
     "ixora": "1.0.10",
     "langs": "2.0.0",
-    "lodash": ">=4.17.21",
+    "lodash": "^4.18.1",
     "most-common-words-by-language": "3.0.10",
     "n": "^8.2.0",
     "prettier": "2.7.1",
