+++ social-analyzer/__init__.py
@@ -1,0 +1,3 @@
+from .analyzer import SocialAnalyzer, main_logic
+
+__all__ = ['SocialAnalyzer', 'main_logic']
