+++ social-analyzer/__main__.py
@@ -1,0 +1,4 @@
+from .analyzer import main_logic
+
+if __name__ == '__main__':
+    main_logic()
