+++ social-analyzer/social-analyzer
@@ -1,0 +1,5 @@
+#!/usr/bin/env python
+from importlib import import_module
+
+if __name__ == '__main__':
+    import_module('social-analyzer').main_logic()
